Rudimentary data point management and exporting
#5
Missing actual scale data or logic

+++ eeprom.ino
@@ -35,11 +35,11 @@ void eepromReadPointer(sint16 PacketIndex) { eeprom.WriteInt16(1020, PacketIndex
 sint16 eepromWritePointer() { return eeprom.ReadInt16(1022); }
 void eepromWritePointer(sint16 PacketIndex) { eeprom.WriteInt16(1022, PacketIndex); }
 // 1024		3072	384 Data samples; 4 byte timestamp, 4 byte value
-TimeAndValue eepromDataSample(int PacketIndex)
+TimeAndValue eepromDataSample(uint16 PacketIndex)
 {
 	return{ eeprom.ReadInt32(1024 + (PacketIndex * 8)), eeprom.ReadInt32(1024 + (PacketIndex * 8) + 4) };
 }
-void eepromDataSample(int PacketIndex, TimeAndValue Data)
+void eepromDataSample(uint16 PacketIndex, TimeAndValue Data)
 {
 	eeprom.WriteInt32(1024 + (PacketIndex * 8), Data.Time);
 	eeprom.WriteInt32(1024 + (PacketIndex * 8) + 4, Data.Value);
